==> 01_merge datasource.ipynb <==
import pandas as pd

#load
df1 = pd.read_csv("data/resale-flat-prices/resale-flat-prices-based-on-approval-date-1990-1999.csv")
df1.head()

#load
df2 = pd.read_csv("data/resale-flat-prices/resale-flat-prices-based-on-approval-date-2000-feb-2012.csv")
df2.head()

#load
df3 = pd.read_csv("data/resale-flat-prices/resale-flat-prices-based-on-registration-date-from-jan-2015-to-dec-2016.csv")
df3.drop('remaining_lease', axis=1, inplace=True)
df3.head()

#load
df4 = pd.read_csv("data/resale-flat-prices/resale-flat-prices-based-on-registration-date-from-jan-2017-onwards.csv")
df4.drop('remaining_lease', axis=1, inplace=True)
df4.head()

#load
df5 = pd.read_csv("data/resale-flat-prices/resale-flat-prices-based-on-registration-date-from-mar-2012-to-dec-2014.csv")
df5.head()

#merge files
merged = pd.concat([df1, df2, df3, df4, df5])
merged.head()

#check
merged.shape[0] - df1.shape[0] - df2.shape[0] - df3.shape[0] - df4.shape[0] - df5.shape[0]

#save file
merged.to_csv("data/resale-flat-prices/merged_prices.csv", index=False)
