==> 02_EDA_clean.ipynb <==
import pandas as pd
#load
df = pd.read_csv("data/resale-flat-prices/merged_prices.csv")
df.head()

#drop duplicates
print(df.shape)
df.drop_duplicates(inplace=True)
df.shape

#check null value and dtypes
df.info()

df.flat_type.value_counts()

df['flat_model'] = df['flat_model'].str.upper()
df.flat_model.value_counts()

#remove discontiued flat models or non-std flats for apple to apple comparison later
df_std = df.where(df.flat_model.isin(['MODEL A', 'IMPROVED', 'NEW GENERATION', 'SIMPLIFIED'
                                      , 'PREMIUM APARTMENT', 'STANDARD', 'APARTMENT', 'MODEL A2', '2-ROOM']))
df_std.dropna(inplace=True)
df_std.flat_type.value_counts()

#treat masionnettes seaprately due to their distinct features and possibly different pricing from executive flat
df_mansionette = df.where(df.flat_model.isin(['MAISONETTE', 'MODEL A-MAISONETTE', 
                                      'IMPROVED-MAISONETTE', 'PREMIUM MAISONETTE']))
df_mansionette.dropna(inplace=True)
df_mansionette.head()

#relassify masionnettes flat type and put it back
df_mansionette['flat_type'] = 'MAISONETTE'
df = pd.concat([df_std, df_mansionette])

#check population size
df.flat_type.value_counts()

#drop flats due to insufficient samples to reduce bias or to be modelled separately
#executive and masionette maintained due to Section 1 Question B
#df_common = df.where(~df.flat_type.isin(["1 ROOM", "2 ROOM"]))
#df_common.dropna(inplace=True)
df_common = df.copy(deep=True)

#check sqm for each flat type
df_common.groupby('flat_type').floor_area_sqm.describe()

df_common['address'] = df_common.block + " " + df_common.street_name
df_big3rm = df_common.where((df_common.flat_type=="3 ROOM") & (df_common.floor_area_sqm>100))
df_big3rm.dropna(inplace=True)
df_big3rm.floor_area_sqm.value_counts()

df_big3rm137 = df_big3rm.where(df_big3rm.floor_area_sqm==137)
df_big3rm137.dropna(inplace=True)
df_big3rm137.address.value_counts()

#compare with the rest of "4 SAGO LANE"
df_compare = df_common.where((df_common.flat_type=="3 ROOM") & (df_common.address=="4 SAGO LANE"))
df_compare.dropna(inplace=True)
df_compare.floor_area_sqm.value_counts()
#likely adjoined flats instead

df_big3rm103 = df_big3rm.where(df_big3rm.floor_area_sqm==103)
df_big3rm103.dropna(inplace=True)
df_big3rm103.address.value_counts()

#compare with the rest of "14 KG ARANG RD"
df_compare = df_common.where((df_common.flat_type=="3 ROOM") & (df_common.address=="14 KG ARANG RD"))
df_compare.dropna(inplace=True)
df_compare.floor_area_sqm.value_counts()
#likely true outlier

#compare with the rest of "12 KG ARANG RD"
df_compare = df_common.where((df_common.flat_type=="3 ROOM") & (df_common.address=="12 KG ARANG RD"))
df_compare.dropna(inplace=True)
df_compare.floor_area_sqm.value_counts()
#likely true outlier

#compare with the rest of "210 NEW UPP CHANGI RD"
df_compare = df_common.where((df.flat_type=="3 ROOM") & (df_common.address=="210 NEW UPP CHANGI RD"))
df_compare.dropna(inplace=True)
df_compare.floor_area_sqm.value_counts()
#likely adjoined flats

#only left 104 and 115 sqm 3-rooms not checked. To drop since only 11 units.
#cleanup the 3 room
df_3rm = df_common.where((df_common.flat_type=="3 ROOM") 
                      & ((df_common.floor_area_sqm<=100)
                          | (df_common.floor_area_sqm==103)
                          | (df_common.floor_area_sqm==107)))
df_3rm.dropna(inplace=True)
df_3rm.floor_area_sqm.describe()
#non-appreciable improvement in statistics

df_common = df_common.where(~(df_common.flat_type=="3 ROOM") )
df_common.dropna(inplace=True)
df_common = pd.concat([df_common, df_3rm])

df_common.town.value_counts()

df_common.storey_range.value_counts()

df_common.head()

#break 'month' into year and month
df_common['year_sold'] = df_common['month'].str[0:4].astype(int)
df_common['mth_sold'] = df_common['month'].str[5:].astype(int)
#check unique values
print(df_common.year_sold.unique())
df_common.mth_sold.unique()

#create age column
df_common['age_sold'] = df_common['year_sold'] - df_common['lease_commence_date']

#cleanup unnecessary columns
df_common.drop(['month', 'address', 'lease_commence_date'], axis=1, inplace=True)
df_common.head()
#not examining resale_price for outliers due to inflation and extraneous factors

df_common.corr()

"""
In March 2017, then-Minister for National Development, Lawrence Wong,
saw fit to respond to the phenomenon of aging, short-lease HDB flats 
being sold at high prices, as reported in Lianhe Zaobao.
2018 rally - LHL addressed issue of expiring leases of HDBS.
"""
df_new = df_common.where((df_common.year_sold<=2016) & (df_common.age_sold>30.0))
df_new.dropna(inplace=True)
df_new.corr()
#age becomes negative correlation

df_new = df_common.where((df_common.year_sold==2016) & (df_common.age_sold>30.0))
df_new.dropna(inplace=True)
df_new.corr()

df_new = df_common.where((df_common.year_sold>2017) & (df_common.age_sold>30.0))
df_new.dropna(inplace=True)
df_new.corr()
#effect of ageing HDBs becomes more pronounced

#save file
df_common.to_csv("data/resale-flat-prices/cleaned_prices.csv", index=False)
