==> 03A_coords_HDB.ipynb <==
import API_Call
import pandas as pd
pd.options.mode.chained_assignment = None
import numpy as np

df = pd.read_csv("data/resale-flat-prices/cleaned_prices.csv")
df.head()

df['block'] = df['block'].str.strip()
df['street_name'] = df['street_name'].str.strip()

df['address'] = df['block'] + ' ' + df['street_name']
addresses = pd.DataFrame(df['address'])
addresses.drop_duplicates(inplace=True)
addresses.reset_index(drop=True, inplace=True)
addresses.head()

len(addresses)

#default latitude/longitude
addresses['lat'] = np.nan
addresses['long'] = np.nan

API_Call.process_df(addresses, "address")

#check any invalid address/request
addresses.isna().sum()

addresses[addresses['lat'].isna()]
#manual random inspection reveals some of these addresses to be outdated and some could not be found in one map api

#merge df
merged = pd.merge(df, addresses, how='left', on='address')
merged.drop('address', axis=1, inplace=True)
merged.head(10)

len(merged)

#address not found, block possibly already en-bloc
merged.dropna(inplace=True)
len(merged)

#save file
merged.to_csv("data/coords/coords_HDB.csv", index=False)
