==> 03B_coords_MRT.ipynb <==
import API_Call
import pandas as pd
pd.options.mode.chained_assignment = None
import numpy as np

df = pd.read_csv("data/coords/mrt_lrt.csv")
df.head()

#station repeated due to different line
print(df.shape)
df.drop_duplicates('Station', keep='last', inplace=True, ignore_index=True)
df.shape

df.drop(['Line'], axis=1, inplace=True)

#default latitude/longitude
df['lat'] = np.nan
df['long'] = np.nan

API_Call.process_df(df, "Station")

#check any invalid address/request
df.isna().sum()

#save file
df.to_csv("data/coords/coords_mrt.csv", index=False)
