==> 03D_coords_Primary.ipynb <==
import API_Call
import pandas as pd

api = "https://developers.onemap.sg/commonapi/search?returnGeom=Y&getAddrDetails=N&searchVal=primary%20school"
list_pri = API_Call.get_list(api)
len(list_pri)

#convert to df for easy viewing and checking
columns = ['School', 'lat', 'long']
df = pd.DataFrame(list_pri, columns=columns)
df.head()

#check any invalid address/request
df.isna().sum()

#save file
df.to_csv("data/coords/coords_primary.csv", index=False)
